+++ main.cu
@@ -1,5 +1,4 @@
 #include "util.hpp"
-#include "util_cuda.h"
 
 #include "streams.hcu"
 #include "graphs.hcu"
@@ -11,13 +10,13 @@ int main(int argc, char** argv) {
     parameters.slots            = read_arg(argc, argv, 2, 1);
     parameters.kernels_per_slot = read_arg(argc, argv, 3, 4);
     parameters.array_size       = 1 << read_arg(argc, argv, 4, 20);
-    parameters.threads          = read_arg(argc, argv, 5, 128);
+    parameters.block_dim        = read_arg(argc, argv, 5, 128);
     parameters.repetitions      = read_arg(argc, argv, 6, 10);
 
     auto total_kernels = parameters.slots*parameters.kernels_per_slot;
     assert(0 == parameters.array_size % total_kernels);
     parameters.array_size_per_kernel = parameters.array_size/total_kernels;
-    parameters.blocks = (parameters.array_size_per_kernel + parameters.threads - 1)/parameters.threads;
+    parameters.grid_dim = num_blocks(parameters.array_size_per_kernel, parameters.block_dim);
 
     std::cout << "array_size          = " << parameters.array_size  << std::endl;
     std::cout << "array_size_per_task = " << parameters.array_size_per_kernel << std::endl;
@@ -25,8 +24,8 @@ int main(int argc, char** argv) {
     std::cout << "repetitions         = " << parameters.repetitions << std::endl;
     std::cout << "slots               = " << parameters.slots << std::endl;
     std::cout << "kernels_per_slot    = " << parameters.kernels_per_slot << std::endl;
-    std::cout << "block_dim           = " << parameters.threads << std::endl;
-    std::cout << "grid_dim            = " << parameters.blocks << std::endl;
+    std::cout << "block_dim           = " << parameters.block_dim << std::endl;
+    std::cout << "grid_dim            = " << parameters.grid_dim << std::endl;
 
     double* xh = malloc_host<double>(parameters.array_size);
     double* yh = malloc_host<double>(parameters.array_size);
